Problem 100
The algorithm/idea is pretty straightforward:

Two trees are the same if and only if their root values are the same, their left subtrees are the same, and their right subtrees are the same. We can use recursion to check the subtrees, which makes the idea very clean to execute.